+++ NetIdMobileSdk/NetIdMobileSdk/api/NetIdConfig.swift
@@ -47,8 +47,6 @@ public struct NetIdConfig {
     public var clientId: String
     /// Redirect URI for your application.  You need to retrieve it from the netID developer portal.
     public var redirectUri: String
-    /// Custom scheme to preceed deep links with - deprecated (will be removed in further version).
-    public var originUrlScheme: String
     /// Additional claims to set.
     public var claims: [String: String]?
     /// Optional configuration for strings to display in the login layer.
@@ -57,11 +55,10 @@ public struct NetIdConfig {
     public var permissionLayerConfig: PermissionLayerConfig?
 
     /// Initialize the SDK. This is the first thing to do.
-    public init(host: String, clientId: String, redirectUri: String, originUrlScheme: String, claims: [String: String]?, loginLayerConfig: LoginLayerConfig?, permissionLayerConfig: PermissionLayerConfig?) {
+    public init(host: String, clientId: String, redirectUri: String, claims: [String: String]?, loginLayerConfig: LoginLayerConfig?, permissionLayerConfig: PermissionLayerConfig?) {
         self.host = host
         self.clientId = clientId
         self.redirectUri = redirectUri
-        self.originUrlScheme = originUrlScheme
         self.claims = claims
         self.loginLayerConfig = loginLayerConfig
         self.permissionLayerConfig = permissionLayerConfig